Small correction to WanaKanaJava.java
Hi!

First of all thanks for your work, I used part of your code in my project (Apache 2.0 licensed) and it works pretty well! I have to thank for your work, and congratulate you too for it.
The thing is that I found a mistake that throws wrong translations when using "_hiraganaToRomaji(String hira)". The problem is that you are concatenating the extra "initial roman char" that "っ" means at the END and not at the BEGINNING of the next romaji-equivalent-kana...

For example: "もっと" is translated to romaji as "Motto (Mo tto)", but the script translates it like "Motot (Mo tot)"
In my suggestion I just changed "romaChar += romaChar.charAt(0);" into "romaChar = romaChar.charAt(0) + romaChar;".

Hope this helps you... 
Adiós!
George Shazkho.